Update TaskList
TaskList does not use streams.

Streams provide a declarative way to write code, making out code more understandable.

The tasklist :: printList() method is changed to create a stream.

printList() is the most relevant method to change as the collectors to list method is natural with the print function together with the "\n".